+++ picodi/integrations/fastapi.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import inspect
 from typing import Any
 
 from picodi._picodi import DependencyCallable, Depends
@@ -13,17 +12,10 @@ __all__ = [
 ]
 
 
-class DependsCallable(Depends):
-    def __call__(self) -> Any:
-        return self
-
-
 class DependsAsyncCallable(Depends):
     async def __call__(self) -> Any:
         return self
 
 
 def Provide(dependency: DependencyCallable, /) -> Any:  # noqa: N802
-    if inspect.iscoroutinefunction(dependency):
-        return DependsAsyncCallable(dependency)
-    return DependsCallable(dependency)
+    return DependsAsyncCallable(dependency)