+++ tests/test_integrations/test_fastapi_integration.py
@@ -3,31 +3,26 @@ from typing import Annotated
 import pytest
 from fastapi import Depends, FastAPI
 from fastapi.exceptions import FastAPIError
-from fastapi.testclient import TestClient
+from starlette.middleware import Middleware
 
-from picodi import (
-    ContextVarScope,
+from picodi import dependency, inject
+from picodi.integrations.fastapi import (
+    PicodiRequestScopeMiddleware,
     Provide,
-    dependency,
-    init_dependencies,
-    inject,
-    shutdown_dependencies,
+    RequestScope,
 )
+
+
+@pytest.fixture()
+def app():
+    return FastAPI(
+        middleware=[Middleware(PicodiRequestScopeMiddleware)],
+    )
 
 
 class MyNumber:
     def __init__(self, number: int):
         self.value = number
-
-
-@pytest.fixture()
-def app():
-    return FastAPI()
-
-
-@pytest.fixture()
-def client(app):
-    return TestClient(app)
 
 
 def test_fastapi_cant_use_provide_as_is(app):
@@ -42,7 +37,7 @@ def test_fastapi_cant_use_provide_as_is(app):
             return {"number": number}
 
 
-def test_resolve_dependency_in_route(app, client):
+async def test_resolve_dependency_in_route(app, asgi_client):
     def get_42() -> MyNumber:
         return MyNumber(42)
 
@@ -51,12 +46,12 @@ def test_resolve_dependency_in_route(app, client):
     def root(number: MyNumber = Depends(Provide(get_42))):
         return {"number": number.value}
 
-    response = client.get("/")
+    response = await asgi_client.get("/")
 
     assert response.json() == {"number": 42}
 
 
-async def test_resolve_dependency_in_route_async(app, client):
+async def test_resolve_dependency_in_route_async(app, asgi_client):
     async def get_42() -> MyNumber:
         return MyNumber(42)
 
@@ -65,12 +60,12 @@ async def test_resolve_dependency_in_route_async(app, client):
     async def root(number: MyNumber = Depends(Provide(get_42))):
         return {"number": number.value}
 
-    response = client.get("/")
+    response = await asgi_client.get("/")
 
     assert response.json() == {"number": 42}
 
 
-def test_resolve_mixed_dependency_in_route(app, client):
+async def test_resolve_mixed_dependency_in_route(app, asgi_client):
     async def get_42() -> MyNumber:
         return MyNumber(42)
 
@@ -84,12 +79,12 @@ def test_resolve_mixed_dependency_in_route(app, client):
     async def root(slug_result: dict[str, int] = Depends(get_by_slug)):
         return slug_result
 
-    response = client.get("/", params={"slug": "meaning-of-life"})
+    response = await asgi_client.get("/", params={"slug": "meaning-of-life"})
 
     assert response.json() == {"meaning-of-life": 42}
 
 
-def test_can_use_depends_only_in_view_not_in_nested_deps(app, client):
+async def test_can_use_provide_in_nested_deps_without_depends(app, asgi_client):
     async def get_42() -> int:
         return 42
 
@@ -114,12 +109,12 @@ def test_can_use_depends_only_in_view_not_in_nested_deps(app, client):
     ):
         return {"nested": string_number.value}
 
-    response = client.get("/")
+    response = await asgi_client.get("/")
 
     assert response.json() == {"nested": "42"}
 
 
-def test_resolve_annotated_dependency(app, client):
+async def test_resolve_annotated_dependency(app, asgi_client):
     def get_42() -> MyNumber:
         return MyNumber(42)
 
@@ -128,66 +123,56 @@ def test_resolve_annotated_dependency(app, client):
     def root(number: Annotated[MyNumber, Depends(Provide(get_42))]):
         return {"number": number.value}
 
-    response = client.get("/")
+    response = await asgi_client.get("/")
 
     assert response.json() == {"number": 42}
 
 
-def test_contextvar_scope_can_be_used_as_request_scope(app, client):
+async def test_middleware_init_and_shutdown_request_scope(app, asgi_client):
+    init_counter = 0
     closing_counter = 0
 
-    @app.middleware("http")
-    async def manage_request_scoped_deps(request, call_next):
-        await init_dependencies(scope_class=ContextVarScope)
-        response = await call_next(request)
-        await shutdown_dependencies(scope_class=ContextVarScope)
-        return response
-
-    @dependency(scope_class=ContextVarScope)
-    def get_42():
-        yield 42
-        nonlocal closing_counter
-        closing_counter += 1
-
-    @app.get("/")
-    @inject
-    def root(
-        number1: int = Depends(Provide(get_42)), number2: int = Depends(Provide(get_42))
-    ):
-        return {"numbers": [number1, number2]}
-
-    response = client.get("/")
-    client.get("/")
-
-    assert closing_counter == 2
-    assert response.json() == {"numbers": [42, 42]}
-
-
-async def test_contextvar_scope_can_be_used_as_request_scope_async(app, client):
-    closing_counter = 0
-
-    @app.middleware("http")
-    async def manage_request_scoped_deps(request, call_next):
-        await init_dependencies(scope_class=ContextVarScope)
-        response = await call_next(request)
-        await shutdown_dependencies(scope_class=ContextVarScope)
-        return response
-
-    @dependency(scope_class=ContextVarScope)
+    @dependency(scope_class=RequestScope)
     async def get_42():
+        nonlocal init_counter
+        init_counter += 1
         yield 42
         nonlocal closing_counter
         closing_counter += 1
 
     @app.get("/")
     @inject
-    async def root(
-        number1: int = Depends(Provide(get_42)), number2: int = Depends(Provide(get_42))
-    ):
-        return {"numbers": [number1, number2]}
+    def root():
+        assert init_counter == 1
+        assert closing_counter == 0
+        return {}
 
-    response = client.get("/")
-    client.get("/")
+    await asgi_client.get("/")
 
-    assert closing_counter == 2
-    assert response.json() == {"numbers": [42, 42]}
+    assert init_counter == 1
+    assert closing_counter == 1
+
+
+async def test_middleware_init_and_shutdown_request_scope_sync(app, asgi_client):
+    init_counter = 0
+    closing_counter = 0
+
+    @dependency(scope_class=RequestScope)
+    def get_42():
+        nonlocal init_counter
+        init_counter += 1
+        yield 42
+        nonlocal closing_counter
+        closing_counter += 1
+
+    @app.get("/")
+    @inject
+    def root():
+        assert init_counter == 1
+        assert closing_counter == 0
+        return {}
+
+    await asgi_client.get("/")
+
+    assert init_counter == 1
+    assert closing_counter == 1